support for graphic system states
it so happens that some of the qt graphics systems are experimental and
don't work well for all (or any) users.

this patch allows specifying the system's state, which then shows up
when listing the available graphics systems with eselect.

the module treats the contents of each graphics system file as its
state.

$ eselect qtgraphicssystem list
Available Qt Graphics Systems:
[1]   native
[2]   opengl (experimental)
[3]   openvg (experimental)
[4]   raster (default) *